Test for navigating around the site
I expect this will eventually become a long test, navigating to all the
different corners of the site.

Probably should be named navigate-anonymous so we can have a similar
test that does the same when logged in as an admin or a public user.

At that point though we need to create a helper module to wrap the
"generic" parts of the navigation in functions so it is easier to do
a bunch of things as different users.

Will have to think on that though.